person ctrl

+++ src/main/java/edu/sjsu/cmpe275/lab2/controller/PersonController.java
@@ -31,18 +31,32 @@ public class PersonController {
     @Autowired
     OrganizationService orgService;
 
+    /**
+     * Create a Person by passing the following parameters
+     *
+     * @param firstname
+     * @param lastname
+     * @param email
+     * @param description
+     * @param street
+     * @param city
+     * @param state
+     * @param zip
+     * @param orgid
+     * @return
+     * @throws Exception
+     */
 
     @RequestMapping(method = RequestMethod.POST)
-    public Person createPerson(@RequestParam(value = "firstname", required = true) String firstname,
-                               @RequestParam(value = "lastname", required = true) String lastname,
-                               @RequestParam(value = "email", required = true) String email,
-                               @RequestParam(value = "description", defaultValue = "") String description,
-                               @RequestParam(value = "street", defaultValue = "") String street,
-                               @RequestParam(value = "city", defaultValue = "") String city,
-                               @RequestParam(value = "state", defaultValue = "state") String state,
-                               @RequestParam(value = "zip", defaultValue = "") String zip,
-                               @RequestParam(value = "organization", defaultValue = "") String orgid,
-                               ModelMap model) throws Exception {
+    public ResponseEntity<Person> createPerson(@RequestParam(value = "firstname", required = true) String firstname,
+                                               @RequestParam(value = "lastname", required = true) String lastname,
+                                               @RequestParam(value = "email", required = true) String email,
+                                               @RequestParam(value = "description", defaultValue = "") String description,
+                                               @RequestParam(value = "street", defaultValue = "") String street,
+                                               @RequestParam(value = "city", defaultValue = "") String city,
+                                               @RequestParam(value = "state", defaultValue = "state") String state,
+                                               @RequestParam(value = "zip", defaultValue = "") String zip,
+                                               @RequestParam(value = "organization", defaultValue = "") String orgid) {
 
 
         Person personObj = new Person();
@@ -53,7 +67,7 @@ public class PersonController {
         if (firstname == null || "".equalsIgnoreCase(firstname)
                 || lastname == null || "".equalsIgnoreCase(lastname)
                 || email == null || "".equalsIgnoreCase(email)) {
-            throw new Exception("required");
+            return new ResponseEntity<Person>(HttpStatus.BAD_REQUEST);
         }
 
         personObj.setFirstname(firstname);
@@ -69,22 +83,42 @@ public class PersonController {
 
         personObj.setFriends(friendList);
 
-        orgObj = orgService.findById(Integer.parseInt(orgid));
-        if (orgObj != null) {
-            personObj.setOrg(orgObj);
-        } else {
-            personObj.setOrg(null);
+
+        /**
+         * if invalid org id then send bad request
+         * */
+
+        if (orgid != null || !"".equalsIgnoreCase(orgid)) {
+            orgObj = orgService.findById(Integer.parseInt(orgid));
+            if (orgObj == null) {
+                return new ResponseEntity<Person>(HttpStatus.BAD_REQUEST);
+            } else {
+                personObj.setOrg(orgObj);
+            }
         }
-        personService.createPerson(personObj);
-        model.addAttribute("person", personObj);
-        return personObj;
+
+        Person createdPerson = personService.createPerson(personObj);
+
+        return new ResponseEntity<Person>(createdPerson, HttpStatus.OK);
     }
+
+    /**
+     * Get a person by Id in HTML format
+     *
+     * @param personId
+     * @param model
+     * @return
+     * @throws EntityNotFound
+     */
 
     @RequestMapping(value = "/{id}",
             method = RequestMethod.GET,
             produces = {"text/html"})
     public String getPersonHTML(@PathVariable(value = "id") int personId, ModelMap model) throws EntityNotFound {
         Person person = personService.getPersonInfo(personId);
+        if (person == null) {
+            throw new EntityNotFound("Person Not Found.");
+        }
         model.addAttribute("person", person);
         return "person";
     }
@@ -104,33 +138,46 @@ public class PersonController {
         }
     }
 
+    /**
+     * Update a person by id
+     *
+     * @param personId
+     * @param email
+     * @param firstname
+     * @param lastname
+     * @param description
+     * @param street
+     * @param city
+     * @param state
+     * @param zip
+     * @param orgid
+     * @return Person in JSON
+     * @throws Exception
+     */
 
     @RequestMapping(value = "/{personId}", method = RequestMethod.POST, produces = "application/json")
     public ResponseEntity<Person> updatePerson(
             @PathVariable(value = "personId") int personId,
             @RequestParam(value = "email", required = true) String email,
-            @RequestParam(value = "firstname", defaultValue = "") String firstname,
-            @RequestParam(value = "lastname", defaultValue = "") String lastname,
+            @RequestParam(value = "firstname", required = true) String firstname,
+            @RequestParam(value = "lastname", required = true) String lastname,
             @RequestParam(value = "description", defaultValue = "") String description,
             @RequestParam(value = "street", defaultValue = "") String street,
             @RequestParam(value = "city", defaultValue = "") String city,
             @RequestParam(value = "state", defaultValue = "") String state,
-            @RequestParam(value = "zip", defaultValue = "") String zip) throws Exception {
+            @RequestParam(value = "zip", defaultValue = "") String zip,
+            @RequestParam(value = "organization", defaultValue = "") String orgid) throws Exception {
 
-        System.out.println("personId:    " + personId);
-        System.out.println("first:  "+ firstname);
         Person person = personService.getPersonInfo(personId);
-
-
         if (person == null) {
             return new ResponseEntity<Person>(HttpStatus.NOT_FOUND);
         }
 
         Address address = person.getAddress();
         person.setEmail(email);
-//        if (firstname != null || "".equalsIgnoreCase(firstname))
-            person.setFirstname(firstname);
 
+        if (firstname != null || "".equalsIgnoreCase(firstname))
+            person.setFirstname(firstname);
         if (lastname != null || "".equalsIgnoreCase(lastname))
             person.setLastname(lastname);
         if (description != null || "".equalsIgnoreCase(description))
@@ -144,26 +191,36 @@ public class PersonController {
         if (zip != null || "".equalsIgnoreCase(zip))
             address.setZip(zip);
 
-
-        int orgId = 2;
-        Organization orgObj = orgService.findById(orgId);
-        if (orgObj == null) {
-            return new ResponseEntity<Person>(HttpStatus.NOT_FOUND);
-        } else {
-            person.setOrg(orgObj);
+        if (orgid != null || "".equalsIgnoreCase(orgid)) {
+            Organization orgObj = orgService.findById(Integer.parseInt(orgid));
+            if (orgObj == null) {
+                return new ResponseEntity<Person>(HttpStatus.NOT_FOUND);
+            } else {
+                person.setOrg(orgObj);
+            }
         }
 
         Person updatedPerson = personService.updatePersonInfo(person);
-//            model.addAttribute("person", updatedPerson);
-//        return "person";
         return new ResponseEntity<Person>(updatedPerson, HttpStatus.OK);
     }
 
 
+    /**
+     * Delete a person
+     *
+     * @param personId
+     * @return
+     * @throws EntityNotFound
+     */
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    public Person deletePerson(@RequestParam(value = "id") int personId) throws EntityNotFound {
-        personService.deletePerson(personId);
-        return new Person();
+    public ResponseEntity<Person> deletePerson(@RequestParam(value = "id") int personId) throws EntityNotFound {
+
+        Person personToDelete = personService.getPersonInfo(personId);
+        if (personToDelete == null)
+            return new ResponseEntity<Person>(HttpStatus.NOT_FOUND);
+
+        Person deletedPerson = personService.deletePerson(personId);
+        return new ResponseEntity<Person>(deletedPerson, HttpStatus.OK);
     }
 
 }
